requirements.txt
mistralai==1.6.0
openai==1.70.0
streamlit==1.44.1
python-dotenv